backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
redis==5.2.0
psycopg[binary]==3.2.13
kafka-python==2.0.2
python-dotenv==1.0.1
PyYAML==6.0.2
stripe==11.1.0
